requirements.txt
feedparser
google-genai
pandas